[Ref]: 프로필 관련 로직 ProfileViewModel로 분리
